WIP - Run Pyre in circleci.
It should copy over the example and find the search-path and typeshed relative to the venv. Open to a better way of doing this.

Libcst [creates](https://github.com/Instagram/LibCST/blob/master/.circleci/config.yml) a venv in a specific location and hardcodes that location in its circleci/.pyre_configuration.

The idea is that users would copy over `.pyre_configuration.example` and hardcode their search path locally. It is git-ignored, so those changes won't be pushed.

TODO: Update the docs.